backend/requirements.txt
fastapi
uvicorn
paho-mqtt
influxdb-client
python-multipart